Fix crash when selecting a module with no assembler selected
Fixed in two ways: first, ensured that the module selector buttons on
new rows get locked (seems a limitation of the factorio API, it doesn't
use the "locked" property when creating an element).

Also added a check when selecting a module that there is an assembler
selected, and return early if not.