Fix various things with import/export
Main changes are wrap import/export functions with pcall. Special case
message when exporting the default-proxy/Anything, as it doesn't work in
blueprints currently (need to figure out a fix for that).

When improting from blueprint, ignore all entities with no module slots.
Before was leaving that index in the config blank, which broke things
assuming differently.

Various other changes to make these broken configs still work. Might
work on a refactor for this related logic, it's kinda messy.