Significant rework, add default and multiple configs
Moves types out to their own file (with a few constructor helpers).

Adds a default config, to be used when no specific row matches.

Adds the data structure to support multiple configs on each line, in
order to allow fallbacks when the main module list isn't compatible
(e.g., main is productivity, then speed that's used when productivity
can't be).