Update on_nth_tick handling
Need to keep track of the registered tick in storage, so it can be
restored when loading (in particular, joining a multiplayer game).

Also fix it not updating the handler when all work was finished (was
returning early from delayed_creation, so never updated the listeners to
remove all, leaving the last one behind every selection).