Ignore prototypes with an obscene number of module slots
Too many slots can cause the gui building code to hang and eventually
crash the game due to making a button for each slot. Ignore prototypes
with huge slot counts (over 512) - that ammount doesn't cause much lag
(for me at least), and anything with that many slots is likely not
user-facing and so doesn't need to be shown by this mod.

A more robust solution would be to adjust the gui display when there are
many slots, and do something more similar to FactoryPlanner.