Add new UI for and rework handling of high module slot counts
Created a new UI to specify amounts of a module, instead of needing to
specify a module for each slot. Allows much better performance and
easier mixing when there are many slots.

Added a preference for when to use the new UI

Tweaked the generation of lookup tables to ignore some entities. Those
with too many slots (currently > 1000) are ignored, and some other
special cases. Added special support to treat SE "grounded" entities the
same as their non-"grounded" counterpart.